pyproject.toml
[tool.poetry]
name = "flame-data-api"
version = "0.1.0"
description = ""
authors = ["Andreas Copan <avcopan@gmail.com>"]
readme = "README.md"
packages = [{include = "flame_data_api"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
psycopg = "^3.1.9"
flask-bcrypt = "^1.0.1"
flask-session = "^0.5.0"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.0.5"
pytest = "^7.4.0"
numpy = "^1.25.2"
qcelemental = "^0.26.0"
pyparsing = "^3.1.1"
more-itertools = "^10.0.0"
scipy = "^1.11.1"
networkx = "^3.1"
ipython = "^8.14.0"
ipywidgets = "^8.1.0"
rdkit = "^2023.3.3"
mako = "^1.2.4"
pyyaml = "^6.0.1"
pandas = "^2.0.3"
py3dmol = "^2.0.3"
psycopg-pool = "^3.1.7"
lxml = "^4.9.3"
notebook = "^7.0.2"
automol = "^2023.8.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
